Toolbar
* Switching to a single toobar class to define tools, not opting for the database route this time.
* Storing namespace in toolbar arrays
* Config array missing namespace
